refactor(util): split tree types/helpers into util/tree.ts
Move TreeElement and its descendants (Directory/File/Documentation
elements + props, the tree-* JSX intrinsics) out of element.ts into a new
tree.ts, leaving element.ts with just the generic Element layer.

Rename and move the tree helpers:
- getElementMap()    → flattenTree()
- getElementPaths()  → getTreePaths()
- resolveElementPath() → resolveTreePath()

Update all importers (modules + docs), split element.test.ts into
element.test.ts + tree.test.ts, and split element.md into element.md +
tree.md.
