+++ modules/extract/PackageExtractor.test.ts
@@ -2,8 +2,8 @@ import { describe, expect, test } from "bun:test";
 import { mkdir, mkdtemp, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import type { DirectoryElement, TreeElement } from "../util/element.js";
 import type { AbsolutePath } from "../util/path.js";
+import type { DirectoryElement, TreeElement } from "../util/tree.js";
 import { DirectoryExtractor } from "./DirectoryExtractor.js";
 import { IndexFileExtractor } from "./IndexFileExtractor.js";
 import { MergingExtractor } from "./MergingExtractor.js";
