+++ modules/ui/tree/TreeSidebar.tsx
@@ -1,6 +1,7 @@
 import type { ReactNode } from "react";
-import { filterElements, type TreeElement } from "../../util/element.js";
+import { filterElements } from "../../util/element.js";
 import type { AbsolutePath } from "../../util/path.js";
+import type { TreeElement } from "../../util/tree.js";
 import { Menu } from "../menu/Menu.js";
 import { MenuItem } from "../menu/MenuItem.js";
 import { matchMenuElement, TreeMenuMapper } from "./TreeMenu.js";
